+++ src/app/(subcategory)/subcategory-ai/components/Form/ConfirmList.tsx
